+++ packages/core/src/extensions/withAbort.ts
@@ -66,7 +66,11 @@ let abortControllers = (
  */
 export let withAbort =
   (
-    strategy: 'last-in-win' | 'first-in-win' | 'manual' = 'last-in-win',
+    strategy:
+      | 'last-in-win'
+      | 'first-in-win'
+      | 'finally'
+      | 'manual' = 'last-in-win',
   ): AssignerExt<AbortExt> =>
   (target) => {
     let recomputed = new WeakSet<Frame>()
@@ -172,6 +176,10 @@ export let withAbort =
             }
             abortSubscription?.unsubscribe()
             rej(error)
+          }
+
+          if (strategy === 'finally') {
+            queueMicrotask(() => thisController.abort('finally'))
           }
         })
         sync = false